@@ -179,7 +179,7 @@ namespace IxMilia.Converters
             return !(matrix1 == matrix2);
         }
 
-        public override bool Equals(object obj)
+        public override bool Equals(object? obj)
         {
             if (obj is Matrix4)
             {
